refactor: 도메인에 해당하는 Exception 폴더 구분
