+++ src/main/java/roomescape/global/error/ExceptionAdvice.java
@@ -4,7 +4,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import roomescape.global.error.exception.BusinessException;
 
 @ControllerAdvice
 public class ExceptionAdvice {
